chore(core): fix tests
