+++ packages/nx/src/native/tests/cache.spec.ts
@@ -3,23 +3,21 @@ import { join } from 'path';
 import { TempFs } from '../../internal-testing-utils/temp-fs';
 import { rmSync } from 'fs';
 import { getDbConnection } from '../../utils/db-connection';
+import { randomBytes } from 'crypto';
 
 describe('Cache', () => {
   let cache: NxCache;
   let tempFs: TempFs;
   let taskDetails: TaskDetails;
 
+  const dbOutputFolder = 'temp-db-cache';
   beforeEach(() => {
     tempFs = new TempFs('cache');
-    rmSync(join(__dirname, 'temp-db'), {
-      recursive: true,
-      force: true,
-    });
 
     const dbConnection = getDbConnection({
-      directory: join(__dirname, 'temp-db'),
+      directory: join(__dirname, dbOutputFolder),
+      dbName: `temp-db-${randomBytes(4).toString('hex')}`,
     });
-
     taskDetails = new TaskDetails(dbConnection);
 
     cache = new NxCache(
@@ -36,6 +34,13 @@ describe('Cache', () => {
         configuration: 'production',
       },
     ]);
+  });
+
+  afterAll(() => {
+    rmSync(join(__dirname, dbOutputFolder), {
+      recursive: true,
+      force: true,
+    });
   });
 
   it('should store results into cache', async () => {
