+++ packages/nx/src/native/tests/task_history.spec.ts
@@ -3,7 +3,9 @@ import { join } from 'path';
 import { TempFs } from '../../internal-testing-utils/temp-fs';
 import { rmSync } from 'fs';
 import { getDbConnection } from '../../utils/db-connection';
+import { randomBytes } from 'crypto';
 
+const dbOutputFolder = 'temp-db-task';
 describe('NxTaskHistory', () => {
   let taskHistory: NxTaskHistory;
   let tempFs: TempFs;
@@ -12,13 +14,9 @@ describe('NxTaskHistory', () => {
   beforeEach(() => {
     tempFs = new TempFs('task-history');
 
-    rmSync(join(__dirname, 'temp-db'), {
-      recursive: true,
-      force: true,
-    });
-
     const dbConnection = getDbConnection({
-      directory: join(__dirname, 'temp-db'),
+      directory: join(__dirname, dbOutputFolder),
+      dbName: `temp-db-${randomBytes(4).toString('hex')}`,
     });
     taskHistory = new NxTaskHistory(dbConnection);
     taskDetails = new TaskDetails(dbConnection);
@@ -38,6 +36,13 @@ describe('NxTaskHistory', () => {
         configuration: 'production',
       },
     ]);
+  });
+
+  afterAll(() => {
+    rmSync(join(__dirname, dbOutputFolder), {
+      recursive: true,
+      force: true,
+    });
   });
 
   it('should record task history', () => {
